[chore] Time 관련 메소드 생성 및 리펙토링
