--- src/main/java/Ness/Backend/global/time/Time.java
@@ -1,13 +1,0 @@
-package Ness.Backend.global.time;
-
-import org.springframework.stereotype.Component;
-
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-
-@Component
-public class Time {
-    public ZonedDateTime getToday(){
-        return ZonedDateTime.now(ZoneId.of("Asia/Seoul"));
-    }
-}
